BSD compile warnings fixed

+++ ex_voper.c
@@ -6,7 +6,7 @@ static char *sccsid = "@(#)ex_voper.c	7.2	10/31/81";
 #include "ex_tty.h"
 #include "ex_vis.h"
 
-#define	blank()		isspace(wcursor[0])
+#define	blank()		isspace((int)wcursor[0])
 #define	forbid(a)	if (a) goto errlab;
 
 static int find(int);
@@ -271,7 +271,7 @@ ein:
 	 */
 	case ',':
 		forbid (lastFKND == 0);
-		c = isupper(lastFKND) ? tolower(lastFKND) : toupper(lastFKND);
+		c = isupper((int)lastFKND) ? tolower((int)lastFKND) : toupper((int)lastFKND);
 		i = lastFCHR;
 		if (vglobp == 0)
 			vglobp = "";
@@ -665,7 +665,7 @@ slerr:
 		if (any(*globp, "^+-."))
 			c = *globp++;
 		i = 0;
-		while (isdigit(*globp))
+		while (isdigit((int)*globp))
 			i = i * 10 + *globp++ - '0';
 		if (any(*globp, "^+-."))
 			c = *globp++;
@@ -818,7 +818,7 @@ int
 wordof(char which, char *wc)
 {
 
-	if (isspace(*wc))
+	if (isspace((int)*wc))
 		return (0);
 	return (!wdkind || wordch(wc) == which);
 }
